Adding comment on DateManager and security for send one time image
